+++ plot/viewer/src/canvases/SketchCanvas.tsx
@@ -451,6 +451,7 @@ function SketchCanvasInner({
           <LayoutControls
             layoutAlgo={layoutAlgo}
             showDirectionSwitch={showDirectionSwitch}
+            doc={doc}
             onLayout={triggerLayout}
             onDirection={treeLayout.layoutInDirection}
           />